Move migration stuff to own mod
